restructured the PostViewHeader component folders
